Code review: fix 4 critical bugs, simplify executor, update docs
Comprehensive review by 4 specialized agents (code-reviewer, silent-failure-hunter,
code-simplifier, comment-analyzer). Net: -341 lines, 4 bugs fixed, 6 doc fixes.

Critical bugs fixed:
- DIVXU.B/W flags computed from un-truncated quotient (N/Z wrong on overflow)
- Timer prescaler off-by-one (0=disabled instead of 0=φ/1, all speeds wrong)
- INC/DEC #2 variants always incremented by 1 instead of 2 (added amount param)
- 78-prefix unknown bit_op returned NOP instead of Unknown (instruction stream desync)

Silent failure fixes:
- Port 7 override changed from u8 to Option<u8> (TestJig adapter port7_value=0 was bypassed)
- Cleaned debug-session comments from decoder

Code simplification (-416 lines):
- SizeParams struct eliminates Byte/Word/Long triplication in executor
- sub_internal unifies exec_sub and exec_cmp (CMP = SUB without writeback)
- shift_value replaces shift_byte/shift_word/shift_long with one generic function
- Timer tick returns [Option<u8>; 5] instead of Vec (no heap allocation per tick)
- Table-driven milestone logging, ITU_BASES constant

Documentation fixes:
- On-chip RAM doc: 64B → 896B (0xFFFB80-0xFFFEFF)
- Boot attempts: "13 entries" → "12 entries" to match table
- Phase 2 log: "Not Started" → "Partially Complete"
- CLAUDE.md: stack guards marked "(planned)"
- 78-prefix MOV.B vs bit-op format fully documented in decoder log

+++ emulator/peripherals/src/wdt.rs
@@ -6,7 +6,7 @@
 
 pub struct Watchdog {
     pub enabled: bool,
-    pub counter: u16,
+    pub counter: u32,
     pub fed: bool,
 }
